Break into geomesa and common packages
